quizzyMind/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
coloredlogs==15.0.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.7
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
durationpy==0.9
faiss-cpu==1.10.0
Faker==37.0.2
fastapi==0.115.11
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langchain==0.3.21
langchain-core==0.3.47
langchain-text-splitters==0.3.7
langsmith==0.3.18
markdown-it-py==3.0.0
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
numpy==2.2.4
oauthlib==3.2.2
onnxruntime==1.21.0
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.15
overrides==7.7.0
packaging==24.2
posthog==3.21.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.1
pypdf==5.4.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlparse==0.5.3
starlette==0.46.1
sympy==1.13.3
tenacity==9.0.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
zipp==3.21.0
zstandard==0.23.0
